// autogpt_platform/frontend/src/app/api/__generated__/models/libraryAgentResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { LibraryAgent } from "./libraryAgent";
import type { Pagination } from "./pagination";

/**
 * Response schema for a list of library agents and pagination info.
 */
export interface LibraryAgentResponse {
  agents: LibraryAgent[];
  pagination: Pagination;
}